refactor: use node-fetch for universal fetch
