+++ package.json
@@ -30,7 +30,8 @@
     "events": "^3.3.0",
     "js-sha256": "^0.9.0",
     "reflect-metadata": "^0.1.13",
-    "uuid": "^8.3.0"
+    "uuid": "^8.3.0",
+    "node-fetch": "^2.6.1"
   },
   "devDependencies": {
     "@types/bn.js": "^5.1.0",
@@ -50,7 +51,6 @@
     "express": "^4.17.1",
     "husky": "^5.1.3",
     "jest": "^26.6.3",
-    "node-fetch": "^2.6.1",
     "npm-run-all": "^4.1.5",
     "prettier": "^2.2.1",
     "release-it": "^14.6.1",
