--- src/transport/http/index.ts
@@ -1,1 +1,0 @@
-export * from './HttpOutboundTransporter'
